// web/src/protocol/generated.ts
export interface ChatPayload {
  sender_id?: string;
  sender_name?: string;
  content: string;
  scope: string;
  time?: number;
  is_system?: boolean;
  message_id?: string;
  room_code?: string;
  game_id?: string;
  server_time?: number;
}